Remove sidecar, types.ts and rename ingestion to referencesAPI
